requirements.txt
fastapi
uvicorn
websockets
twilio
google-genai
python-dotenv
pydantic-settings
